The Promise.allof was not working when in a a react
It appears that if there was no Promise or all the Promises
were Kept that the allof used for validation was never being
Kept, so in these cases don't even bother with an additional
Promise and return a new one directly if all the results are
known at the time.  This probably represents an optimisation.